requirements.txt
python-telegram-bot==21.11.1
openai>=1.109.0
pydantic>=2.10.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
APScheduler>=3.10.4
psutil>=5.9.8
send2trash>=1.8.3
python-docx>=1.1.2
pypdf>=5.5.0
pyautogui>=0.9.54
Pillow>=11.0.0
comtypes>=1.4.8
pycaw>=20240210
requests>=2.32.3
pydub>=0.25.1
faster-whisper>=1.1.1
pytest>=8.3.4
